requirements_trepflow.txt
absl-py==2.0.0
accelerate==0.31.0
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
annotated-types==0.7.0
anyio==4.0.0
appdirs==1.4.4
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
black==23.10.1
blessed==1.20.0
boto3==1.28.78
botocore==1.31.78
brotlipy==0.7.0
cachetools==5.3.2
cattrs==23.1.2
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.1.1
correlation==0.0.0
correlation-cuda==0.0.0
cppimport==22.8.2
croniter==1.3.15
cupy-cuda12x==13.1.0
cycler==0.12.1
dateutils==0.6.12
debugpy==1.8.0
decorator==5.1.1
deepdiff==6.6.1
docker-pycreds==0.4.0
docutils==0.20.1
einops==0.6.1
esbonio==0.16.2
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.88.0
fastjsonschema==2.19.1
fastrlock==0.8.2
filelock==3.12.3
flash-attn==2.5.9.post1
flatbuffers==24.3.7
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.0
gitdb==4.0.10
GitPython==3.1.35
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.2
h11==0.14.0
h5py==3.10.0
hachibee-sphinx-theme==0.2.5
huggingface-hub==0.23.4
humanfriendly==10.0
imageio==2.32.0
imagesize==1.4.1
iniconfig==2.0.0
inquirer==3.1.3
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab-widgets==3.0.9
kaleido==0.2.1
kiwisolver==1.4.5
lazy_loader==0.3
lightning==1.9.5
lightning-cloud==0.5.50
lightning-utilities==0.9.0
lsprotocol==2023.0.0b1
Mako==1.3.5
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.6
mkl-service==2.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numpydoc==1.6.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-modelopt==0.13.0
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.16.1
onnx-graphsurgeon==0.5.2
onnxruntime-gpu==1.18.0
onnxruntime_extensions==0.11.0
onnxscript==0.1.0.dev20240624
opencv-python==4.8.0.76
ordered-set==4.1.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pexpect==4.9.0
Pillow==9.4.0
platformdirs==3.11.0
plotly==5.16.1
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.12.0
pydantic==1.10.17
pydantic_core==2.18.4
pygls==1.1.2
Pygments==2.16.1
PyJWT==2.8.0
pynvml==11.5.0
pyparsing==3.1.1
pypng==0.20220715.0
pyspellchecker==0.7.2
pytest==7.4.4
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.3.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
quadtree-attention-package==0.0.0
readchar==4.0.5
referencing==0.33.0
regex==2024.5.15
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.18.0
rsa==4.9
s3transfer==0.7.0
safetensors==0.4.3
scikit-image==0.22.0
scipy==1.11.2
sentry-sdk==1.30.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
spatial_correlation_sampler==0.5.0
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0rc4
sphinx-sizzle-theme==0.1.3
sphinx-theme-pd==0.8.3
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
stack-data==0.6.3
starlette==0.22.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorrt==10.0.1
tensorrt-cu12==10.1.0
tensorrt-cu12-bindings==10.1.0
tensorrt-cu12-libs==10.1.0
tifffile==2023.9.26
timm==0.9.7
tokenize-rt==5.2.0
tokenizers==0.19.1
tomli==2.0.1
torchmetrics==1.1.1
tornado==6.4
tqdm==4.66.1
traitlets==5.13.0
transformers==4.41.2
triton==2.3.1
typeguard==4.1.5
types-python-dateutil==2.8.19.14
typing_extensions==4.10.0
tzdata==2023.3
uvicorn==0.24.0
wandb==0.15.10
wcwidth==0.2.9
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==4.0.9
yacs==0.1.8
yarl==1.9.2
